Refactor getFile/GetPage functions
