+++ src/page/index.ts
@@ -1,3 +1,9 @@
 export { getPageInfo } from './getPageInfo';
 export { makePageElement } from './makePageElement';
 export { getPageComponents } from './getPageComponents';
+export {
+  getSinglePageFile,
+  getSinglePageFileIfExists,
+  getPageFile,
+  getPageFileIfExists,
+} from './getPageFile';
